requirements.txt
# On Ubuntu, install the following packages:
# - python3
# - python3-venv
#
# And then the virtual environment can be created by running:
#
# python3 -m venv .venv
# .venv/bin/pip install --upgrade pip
# .venv/bin/pip install -r requirements.txt
#
# Then use .venv/bin/python to run the various training scripts.

jupyter
matplotlib>=3.6.2
torch>=1.13.0
torchvision~=0.14.1
numpy>=1.16.4
torchsummary
pandas
albumentations
seaborn>=0.12.2
tqdm~=4.65.0

Pillow~=9.4.0